test(web): extract install_default_stubs into tests/web/conftest.py (#879)
PR #878 review (suggestion 1) flagged that ``_install_default_stubs``
was duplicated verbatim across 5+ spec files; the count had grown to 7
on main. Move the canonical 7-route stub helper to
``tests/web/conftest.py`` as ``install_default_stubs(page)`` next to
the existing ``mm_web_url`` fixture.

Pre-collapse parity grep across all 8 caller sites found two drift
points worth pinning in the PR rather than papering over:

1. **``test_sources_reindex_retry.py``** uses a fundamentally
   different stub set (``sources?**`` with query-string suffix, no
   ``namespaces`` / ``privacy/patterns`` routes, adds
   ``memory-dirs/status``). Out of #879 scope — leave its local
   helper alone. The shared helper is for the boot-fetch family
   (``ui-mode`` / ``sources`` / ``namespaces`` / ``stats`` /
   ``privacy/patterns``), not for memory-dirs scope.
2. **``test_tag_filter_mutation.py``** registers an additional
   ``**/api/search?**`` empty default that specs override with a
   capturing handler. Keep this as a thin local wrapper that
   delegates to the shared helper and adds the search route — call
   sites stay identical and last-route-wins ordering is preserved.

Six specs migrated wholesale to the shared helper:
- ``test_redaction_blocked_retry.py``
- ``test_context_gateway_overview.py``
- ``test_context_gateway_lists.py``
- ``test_context_gateway_sync_all.py``
- ``test_settings_hooks_sync_confirm.py``
- ``test_settings_hooks_resolve_inline.py``

Net: -97 LOC (203 deleted, 106 added). All 43 browser tests still
collect; ruff check + format clean. No assertion changes, no test
rewiring, no production code changed.
